experiments/langchain_lab/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
faiss-cpu==1.8.0
flake8==7.0.0
Flask==3.0.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
google-search-results==2.4.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
pinecone-client==3.1.0
platformdirs==4.2.0
protobuf==4.25.3
pyarrow==15.0.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pyflakes==3.2.0
Pygments==2.17.2
pypdf==4.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
streamlit==1.31.1
streamlit-chat==0.1.1
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
